#132 曲編集画面にインターネットランキングを表示する #133 オンラインAPIとの通信のタイムアウト、リトライ制御 #108 DJポイント機能追加 冗長なメソッド開始・終了ログをコメントアウト

+++ app/src/main/java/net/in/ahr/dbms/data/network/api/dto/MusicResultDBHRDto.java
@@ -30,7 +30,7 @@ public class MusicResultDBHRDto {
     private String clearLamp_RH;
     private String clearLamp_LH;
     private String myDifficult;
-    private Integer djPoint;
+    private Double djPoint;
     private Double clearProgressRate;
     private java.util.Date lastPlayDate;
     private java.util.Date lastUpdateDate;
@@ -205,11 +205,11 @@ public class MusicResultDBHRDto {
         this.myDifficult = myDifficult;
     }
 
-    public Integer getDjPoint() {
+    public Double getDjPoint() {
         return djPoint;
     }
 
-    public void setDjPoint(Integer djPoint) {
+    public void setDjPoint(Double djPoint) {
         this.djPoint = djPoint;
     }
